ref:投稿削除時に、storageの中身とサブコレクションを削除する
